test(sema): move natspec tests to analysis
